backend/pyproject.toml
[project]
name = "smart-library"
version = "0.1.0"
description = "FastAPI backend for Smart Library web app"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "psycopg[binary]>=3.1",
    "alembic>=1.13",
    "pydantic[email]>=2.7",
    "pydantic-settings>=2.2",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0",
    "pyjwt[crypto]>=2.8",
    "cryptography>=42.0",
    "reportlab>=4.0",
    "qrcode[pil]>=7.4",
    "python-dateutil>=2.9",
    "loguru>=0.7",
    "aiosqlite>=0.19"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["resources*", "storage*"]
